refactor: update cash in/out flows
