+++ packages/cryptoplease/lib/app/screens/authenticated/cash_out/cash_out_screen.dart
@@ -1,4 +1,3 @@
-import 'package:cryptoplease/app/components/partner_button.dart';
 import 'package:cryptoplease/config.dart';
 import 'package:cryptoplease/core/presentation/utils.dart';
 import 'package:cryptoplease/gen/assets.gen.dart';
@@ -6,6 +5,7 @@ import 'package:cryptoplease/l10n/l10n.dart';
 import 'package:cryptoplease/ui/app_bar.dart';
 import 'package:cryptoplease/ui/colors.dart';
 import 'package:cryptoplease/ui/content_padding.dart';
+import 'package:cryptoplease/ui/partner_button.dart';
 import 'package:cryptoplease/ui/theme.dart';
 import 'package:flutter/material.dart';
