+++ packages/cryptoplease/lib/app/screens/authenticated/routes.dart
@@ -10,7 +10,7 @@ import 'package:cryptoplease/app/screens/authenticated/profile/menu_screen.dart'
 import 'package:cryptoplease/app/screens/authenticated/profile/profile_screen.dart';
 import 'package:cryptoplease/app/screens/authenticated/receive_flow/routes.dart';
 import 'package:cryptoplease/app/screens/authenticated/wallet_flow/wallet_flow_screen.dart';
-import 'package:cryptoplease/features/add_funds/presentation/routes.dart';
+import 'package:cryptoplease/features/add_funds/presentation/add_funds_screen.dart';
 import 'package:cryptoplease/features/app_lock/presentation/routes.dart';
 import 'package:cryptoplease/features/backup_phrase/presentation/routes.dart';
 import 'package:cryptoplease/features/incoming_split_key_payments/presentation/routes.dart';
@@ -38,7 +38,7 @@ const authenticatedFlowRoutes = AutoRoute<dynamic>(
     ...iskpRoutes,
     AutoRoute<QrScannerRequest>(page: QrScannerScreen),
     ...receiveFlowRoutes,
-    ...addFundsRoutes,
+    AutoRoute<void>(page: AddFundsScreen, fullscreenDialog: true),
     AutoRoute<void>(page: CashOutScreen, fullscreenDialog: true),
     appLockSetupFlowRoutes,
     AutoRoute<void>(page: ProfileScreen, fullscreenDialog: true),
