--- packages/cryptoplease/lib/features/add_funds/presentation/routes.dart
@@ -1,6 +1,0 @@
-import 'package:auto_route/auto_route.dart';
-import 'package:cryptoplease/features/add_funds/presentation/add_funds_screen.dart';
-
-const addFundsRoutes = [
-  AutoRoute<void>(page: AddFundsScreen, fullscreenDialog: true),
-];
